fix: restore promotion validation
